+++ qiskit/providers/ibmq/api_v2/clients/account.py
@@ -58,13 +58,16 @@ class AccountClient(BaseClient):
 
     # Backend-related public functions.
 
-    def list_backends(self) -> List[Dict[str, Any]]:
+    def list_backends(self, timeout: Optional[float] = None) -> List[Dict[str, Any]]:
         """Return the list of backends.
+
+        Args:
+            timeout (float or None): number of seconds to wait for the request.
 
         Returns:
             list[dict]: a list of backends.
         """
-        return self.client_api.backends()
+        return self.client_api.backends(timeout=timeout)
 
     def backend_status(self, backend_name: str) -> Dict[str, Any]:
         """Return the status of a backend.
@@ -367,9 +370,9 @@ class AccountClient(BaseClient):
         # pylint: disable=missing-docstring
         return self.backend_pulse_defaults(backend)
 
-    def available_backends(self):
+    def available_backends(self, timeout=None):
         # pylint: disable=missing-docstring
-        return self.list_backends()
+        return self.list_backends(timeout=timeout)
 
     def get_job(self, id_job, exclude_fields=None, include_fields=None):
         # pylint: disable=missing-docstring